Start implementation of generic reduce algo
